+++ examples/pypi.rs
@@ -5,7 +5,8 @@ fn main() {
     let pkg_name = "filprofiler";
     let current_version = "2022.1.0";
 
-    let informer = UpdateInformer::new(PyPI, pkg_name, current_version, Duration::default());
+    let informer =
+        UpdateInformer::new(PyPI, pkg_name, current_version).interval(Duration::default());
 
     if let Ok(Some(new_version)) = informer.check_version() {
         println!("A new release of {pkg_name} is available: v{current_version} -> {new_version}");